requirements.txt
# CSC110 Project: Python libraries we'll be using for the project.

# Testing and code checking
python-ta

# Graphics and data visualization
plotly.express
pandas

# Randomisation
random

# Data analysis
csv